# docs/openapi/with_auth.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cohere Toolkit API",
    "description": "This is the API for the Open Source Cohere Toolkit",
    "version": "v1.1.6"
  },
  "paths": {
...
    "/v1/organizations": {
      "post": {
        "tags": ["organization"],
        "summary": "Create Organization",
        "description": "Create a new organization.",
        "operationId": "create_organization_v1_organizations_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrganization" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Organization" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["organization"],
        "summary": "List Organizations",
        "description": "List all available organizations.",
        "operationId": "list_organizations_v1_organizations_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Organization" },
                  "title": "Response List Organizations V1 Organizations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/organizations/{org_id}": {
      "put": {
        "tags": ["organization"],
        "summary": "Update Organization",
        "description": "Update organization by ID.",
        "operationId": "update_organization_v1_organizations__org_id__put",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization ID",
              "description": "Organization ID for the organization in question"
            },
            "description": "Organization ID for the organization in question"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateOrganization" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Organization" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["organization"],
        "summary": "Get Organization",
        "description": "Get a organization by ID.",
        "operationId": "get_organization_v1_organizations__org_id__get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization ID",
              "description": "Organization ID for the organization in question"
            },
            "description": "Organization ID for the organization in question"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Organization" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["organization"],
        "summary": "Delete Organization",
        "description": "Delete a organization by ID.",
        "operationId": "delete_organization_v1_organizations__org_id__delete",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization ID",
              "description": "Organization ID for the organization in question"
            },
            "description": "Organization ID for the organization in question"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteOrganization" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/organizations/{org_id}/users": {
      "get": {
        "tags": ["organization"],
        "summary": "Get Organization Users",
        "description": "Get organization users by ID.",
        "operationId": "get_organization_users_v1_organizations__org_id__users_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization ID",
              "description": "Organization ID for the organization in question"
            },
            "description": "Organization ID for the organization in question"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/backend__schemas__user__User"
                  },
                  "title": "Response Get Organization Users V1 Organizations  Org Id  Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrganization": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of the organization"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "CreateOrganization",
        "description": "Request to create an organization"
      },
...
      "DeleteOrganization": {
        "properties": {},
        "type": "object",
        "title": "DeleteOrganization",
        "description": "Response when deleting organization"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Organization": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of the organization"
          },
          "id": {
            "type": "string",
            "title": "ID",
            "description": "Unique identifier of the organization"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Timestamp",
            "description": "When organization was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At Timestamp",
            "description": "When organization was updated"
          }
        },
        "type": "object",
        "required": ["name", "id", "created_at", "updated_at"],
        "title": "Organization",
        "description": "Schema for an organization"
      },
...
      "UpdateOrganization": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name",
            "description": "Name of the organization"
          }
        },
        "type": "object",
        "title": "UpdateOrganization",
        "description": "Request to update an organization"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
      "backend__schemas__user__User": {
        "properties": {
          "fullname": {
            "type": "string",
            "title": "Full Name",
            "description": "User's Full Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email",
            "description": "User's email address"
          },
          "id": { "type": "string", "title": "ID", "description": "" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Timestamp",
            "description": "When the user was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At Timestamp",
            "description": "When the user was updated"
          }
        },
        "type": "object",
        "required": ["fullname", "id", "created_at", "updated_at"],
        "title": "User",
        "description": "User schema"
      }
    },
    "securitySchemes": { "HTTPBasic": { "type": "http", "scheme": "basic" } }
  }
}